do not include indexes/FKs if the underlying table is not present
This will typically happen only if a filter has been specified
for the "copy" command. Other database objects are not affected
as they are either children of the filtered node.

This still leaves the scenario when the dependency cannot be
inferred directly (e.g. a view being used in a package) but
that will require redesign to be handled properly.